add showonsignup to projects
